Clean up int -> long ID conversion
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonWebsite/trunk@83124 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Model/src/java/org/apidb/apicommon/controller/EditCommentAction.java
@@ -27,7 +27,7 @@ public class EditCommentAction extends CommentAction {
 
         // get the comments for the (project_id, stable_id) tuple
         EditCommentForm editForm = (EditCommentForm) form;
-        Comment comment = factory.getComment(Integer.valueOf(editForm.getCommentId()).intValue());
+        Comment comment = factory.getComment(Long.valueOf(editForm.getCommentId()).longValue());
 
         // set for the forwarding page
         String stableId  = editForm.getStableId();
@@ -50,9 +50,9 @@ public class EditCommentAction extends CommentAction {
         ExternalDatabase[] edbs =  comment.getExternalDbs();
         ExternalDatabase edb = edbs[0];
 
-        int[] categoryIds = comment.getTargetCategoryIds();
+        long[] categoryIds = comment.getTargetCategoryIds();
         String categoryString = "";
-        if(categoryIds != null && categoryIds.length > 0) {
+        if (categoryIds != null && categoryIds.length > 0) {
           for(int i = 0; i < categoryIds.length; i++) {
            categoryString += "&targetCategory=" + categoryIds[i]; 
           }
